+++ tests/test_ec2_metadata.py
@@ -12,7 +12,7 @@ def clear_it():
 @pytest.fixture(autouse=True)
 def em_requests_mock(requests_mock):
     requests_mock.put(
-        "http://169.254.169.254/latest/api/token",
+        "http://[fd00:ec2::254]/latest/api/token",
         headers={"X-aws-ec2-metadata-token-ttl-seconds": str(TOKEN_TTL_SECONDS)},
         text="example-token",
     )
@@ -26,11 +26,24 @@ example_mac = "00:11:22:33:44:55"
 
 
 def test_custom_session(em_requests_mock):
-    EC2Metadata(session=requests.Session())
+    em = EC2Metadata(session=requests.Session())
     em_requests_mock.get(
+        "http://[fd00:ec2::254]/latest/meta-data/ami-id", text="ami-12345678"
+    )
+    assert em.ami_id == "ami-12345678"
+
+
+def test_ipv4(requests_mock):
+    em = EC2Metadata(ipv4=True)
+    requests_mock.put(
+        "http://169.254.169.254/latest/api/token",
+        headers={"X-aws-ec2-metadata-token-ttl-seconds": str(TOKEN_TTL_SECONDS)},
+        text="example-token",
+    )
+    requests_mock.get(
         "http://169.254.169.254/latest/meta-data/ami-id", text="ami-12345678"
     )
-    assert ec2_metadata.ami_id == "ami-12345678"
+    assert em.ami_id == "ami-12345678"
 
 
 def add_identity_doc_response(em_requests_mock, overrides=None):
@@ -48,7 +61,7 @@ def add_identity_doc_response(em_requests_mock, overrides=None):
     if overrides:
         identity_doc.update(overrides)
     em_requests_mock.get(
-        "http://169.254.169.254/latest/dynamic/instance-identity/document",
+        "http://[fd00:ec2::254]/latest/dynamic/instance-identity/document",
         json=identity_doc,
     )
     return identity_doc
@@ -61,7 +74,7 @@ def test_account_id(em_requests_mock):
 
 def test_account_id_token_error(requests_mock):
     requests_mock.put(
-        "http://169.254.169.254/latest/api/token",
+        "http://[fd00:ec2::254]/latest/api/token",
         headers={"X-aws-ec2-metadata-token-ttl-seconds": str(TOKEN_TTL_SECONDS)},
         status_code=500,
     )
@@ -71,7 +84,7 @@ def test_account_id_token_error(requests_mock):
 
 def test_account_id_error(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/dynamic/instance-identity/document",
+        "http://[fd00:ec2::254]/latest/dynamic/instance-identity/document",
         status_code=500,
     )
     with pytest.raises(requests.exceptions.HTTPError):
@@ -80,31 +93,31 @@ def test_account_id_error(em_requests_mock):
 
 def test_ami_id(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/ami-id", text="ami-12345678"
+        "http://[fd00:ec2::254]/latest/meta-data/ami-id", text="ami-12345678"
     )
     assert ec2_metadata.ami_id == "ami-12345678"
 
 
 def test_ami_id_cached(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/ami-id", text="ami-12345678"
+        "http://[fd00:ec2::254]/latest/meta-data/ami-id", text="ami-12345678"
     )
     ec2_metadata.ami_id
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/ami-id", status_code=500
+        "http://[fd00:ec2::254]/latest/meta-data/ami-id", status_code=500
     )
     ec2_metadata.ami_id  # no error
 
 
 def test_ami_id_cached_cleared(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/ami-id", text="ami-12345678"
+        "http://[fd00:ec2::254]/latest/meta-data/ami-id", text="ami-12345678"
     )
     ec2_metadata.ami_id
 
     ec2_metadata.clear_all()
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/ami-id", status_code=500
+        "http://[fd00:ec2::254]/latest/meta-data/ami-id", status_code=500
     )
 
     with pytest.raises(requests.exceptions.HTTPError):
@@ -113,21 +126,21 @@ def test_ami_id_cached_cleared(em_requests_mock):
 
 def test_ami_launch_index(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/ami-launch-index", text="0"
+        "http://[fd00:ec2::254]/latest/meta-data/ami-launch-index", text="0"
     )
     assert ec2_metadata.ami_launch_index == 0
 
 
 def test_ami_manifest_path(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/ami-manifest-path", text="(unknown)"
+        "http://[fd00:ec2::254]/latest/meta-data/ami-manifest-path", text="(unknown)"
     )
     assert ec2_metadata.ami_manifest_path == "(unknown)"
 
 
 def test_availability_zone(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/placement/availability-zone",
+        "http://[fd00:ec2::254]/latest/meta-data/placement/availability-zone",
         text="eu-west-1a",
     )
     assert ec2_metadata.availability_zone == "eu-west-1a"
@@ -135,27 +148,27 @@ def test_availability_zone(em_requests_mock):
 
 def test_availability_zone_id(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/placement/availability-zone-id",
+        "http://[fd00:ec2::254]/latest/meta-data/placement/availability-zone-id",
         text="use1-az6",
     )
     assert ec2_metadata.availability_zone_id == "use1-az6"
 
 
 def test_iam_info(em_requests_mock):
-    em_requests_mock.get("http://169.254.169.254/latest/meta-data/iam/info", text="{}")
+    em_requests_mock.get("http://[fd00:ec2::254]/latest/meta-data/iam/info", text="{}")
     assert ec2_metadata.iam_info == {}
 
 
 def test_iam_info_none(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/iam/info", status_code=404
+        "http://[fd00:ec2::254]/latest/meta-data/iam/info", status_code=404
     )
     assert ec2_metadata.iam_info is None
 
 
 def test_iam_info_unexpected(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/iam/info", status_code=500
+        "http://[fd00:ec2::254]/latest/meta-data/iam/info", status_code=500
     )
     with pytest.raises(requests.exceptions.HTTPError):
         ec2_metadata.iam_info
@@ -163,14 +176,14 @@ def test_iam_info_unexpected(em_requests_mock):
 
 def test_instance_action(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/instance-action", text="none"
+        "http://[fd00:ec2::254]/latest/meta-data/instance-action", text="none"
     )
     assert ec2_metadata.instance_action == "none"
 
 
 def test_instance_id(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/instance-id", text="i-12345678"
+        "http://[fd00:ec2::254]/latest/meta-data/instance-id", text="i-12345678"
     )
     assert ec2_metadata.instance_id == "i-12345678"
 
@@ -182,7 +195,7 @@ def test_instance_identity(em_requests_mock):
 
 def test_instance_profile_arn(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/iam/info",
+        "http://[fd00:ec2::254]/latest/meta-data/iam/info",
         text='{"InstanceProfileArn": "arn:foobar"}',
     )
     assert ec2_metadata.instance_profile_arn == "arn:foobar"
@@ -190,14 +203,14 @@ def test_instance_profile_arn(em_requests_mock):
 
 def test_instance_profile_arn_none(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/iam/info", status_code=404
+        "http://[fd00:ec2::254]/latest/meta-data/iam/info", status_code=404
     )
     assert ec2_metadata.instance_profile_arn is None
 
 
 def test_instance_profile_id(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/iam/info",
+        "http://[fd00:ec2::254]/latest/meta-data/iam/info",
         text='{"InstanceProfileId": "some-id"}',
     )
     assert ec2_metadata.instance_profile_id == "some-id"
@@ -205,42 +218,42 @@ def test_instance_profile_id(em_requests_mock):
 
 def test_instance_profile_id_none(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/iam/info", status_code=404
+        "http://[fd00:ec2::254]/latest/meta-data/iam/info", status_code=404
     )
     assert ec2_metadata.instance_profile_id is None
 
 
 def test_instance_type(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/instance-type", text="t2.nano"
+        "http://[fd00:ec2::254]/latest/meta-data/instance-type", text="t2.nano"
     )
     assert ec2_metadata.instance_type == "t2.nano"
 
 
 def test_kernel_id(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/kernel-id", text="aki-dc9ed9af"
+        "http://[fd00:ec2::254]/latest/meta-data/kernel-id", text="aki-dc9ed9af"
     )
     assert ec2_metadata.kernel_id == "aki-dc9ed9af"
 
 
 def test_kernel_id_none(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/kernel-id", status_code=404
+        "http://[fd00:ec2::254]/latest/meta-data/kernel-id", status_code=404
     )
     assert ec2_metadata.kernel_id is None
 
 
 def test_mac(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/mac", text=example_mac
+        "http://[fd00:ec2::254]/latest/meta-data/mac", text=example_mac
     )
     assert ec2_metadata.mac == example_mac
 
 
 def test_network_interfaces(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/network/interfaces/macs/",
+        "http://[fd00:ec2::254]/latest/meta-data/network/interfaces/macs/",
         text=example_mac + "/",
     )
     assert ec2_metadata.network_interfaces == {
@@ -250,7 +263,7 @@ def test_network_interfaces(em_requests_mock):
 
 def test_private_hostname(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/local-hostname",
+        "http://[fd00:ec2::254]/latest/meta-data/local-hostname",
         text="ip-172-30-0-0.eu-west-1.compute.internal",
     )
     assert ec2_metadata.private_hostname == "ip-172-30-0-0.eu-west-1.compute.internal"
@@ -258,14 +271,14 @@ def test_private_hostname(em_requests_mock):
 
 def test_private_ipv4(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/local-ipv4", text="172.30.0.0"
+        "http://[fd00:ec2::254]/latest/meta-data/local-ipv4", text="172.30.0.0"
     )
     assert ec2_metadata.private_ipv4 == "172.30.0.0"
 
 
 def test_public_hostname(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/public-hostname",
+        "http://[fd00:ec2::254]/latest/meta-data/public-hostname",
         text="ec2-1-2-3-4.compute-1.amazonaws.com",
     )
     assert ec2_metadata.public_hostname == "ec2-1-2-3-4.compute-1.amazonaws.com"
@@ -273,21 +286,21 @@ def test_public_hostname(em_requests_mock):
 
 def test_public_hostname_none(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/public-hostname", status_code=404
+        "http://[fd00:ec2::254]/latest/meta-data/public-hostname", status_code=404
     )
     assert ec2_metadata.public_hostname is None
 
 
 def test_public_ipv4(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/public-ipv4", text="1.2.3.4"
+        "http://[fd00:ec2::254]/latest/meta-data/public-ipv4", text="1.2.3.4"
     )
     assert ec2_metadata.public_ipv4 == "1.2.3.4"
 
 
 def test_public_ipv4_none(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/public-ipv4", status_code=404
+        "http://[fd00:ec2::254]/latest/meta-data/public-ipv4", status_code=404
     )
     assert ec2_metadata.public_ipv4 is None
 
@@ -299,7 +312,7 @@ def test_region(em_requests_mock):
 
 def test_reservation_id(em_requests_mock):
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/reservation-id",
+        "http://[fd00:ec2::254]/latest/meta-data/reservation-id",
         text="r-12345678901234567",
     )
     assert ec2_metadata.reservation_id == "r-12345678901234567"
@@ -308,7 +321,7 @@ def test_reservation_id(em_requests_mock):
 def test_security_groups_single(em_requests_mock):
     # most common case: a single SG
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/security-groups",
+        "http://[fd00:ec2::254]/latest/meta-data/security-groups",
         text="security-group-one",
     )
     assert ec2_metadata.security_groups == ["security-group-one"]
@@ -317,7 +330,7 @@ def test_security_groups_single(em_requests_mock):
 def test_security_groups_two(em_requests_mock):
     # another common case: multiple SGs
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/security-groups",
+        "http://[fd00:ec2::254]/latest/meta-data/security-groups",
         text="security-group-one\nsecurity-group-2",
     )
     assert ec2_metadata.security_groups == ["security-group-one", "security-group-2"]
@@ -327,18 +340,18 @@ def test_security_groups_emptystring(em_requests_mock):
     # Check '' too. Can't create an instance without a SG on EC2 but we should
     # safely handle it, perhaps it's possible in e.g. OpenStack.
     em_requests_mock.get(
-        "http://169.254.169.254/latest/meta-data/security-groups", text=""
+        "http://[fd00:ec2::254]/latest/meta-data/security-groups", text=""
     )
     assert ec2_metadata.security_groups == []
 
 
 def test_user_data_none(em_requests_mock):
-    em_requests_mock.get("http://169.254.169.254/latest/user-data/", status_code=404)
+    em_requests_mock.get("http://[fd00:ec2::254]/latest/user-data/", status_code=404)
     assert ec2_metadata.user_data is None
 
 
 def test_user_data_something(em_requests_mock):
-    em_requests_mock.get("http://169.254.169.254/latest/user-data/", content=b"foobar")
+    em_requests_mock.get("http://[fd00:ec2::254]/latest/user-data/", content=b"foobar")
     assert ec2_metadata.user_data == b"foobar"
 
 
@@ -347,7 +360,7 @@ def test_user_data_something(em_requests_mock):
 
 def add_interface_response(em_requests_mock, url, text="", **kwargs):
     full_url = (
-        "http://169.254.169.254/latest/meta-data/network/interfaces/macs/"
+        "http://[fd00:ec2::254]/latest/meta-data/network/interfaces/macs/"
         + example_mac
         + url
     )